[NEW] added react-router-dom + dismiss auto alerts at 5s
