+++ src/components/NotificationAlert.tsx
@@ -1,6 +1,15 @@
 import { NotificationProps } from '../interfaces';
 
 const NotificationAlert = ({ error, message }: NotificationProps) => {
+  const interval = setInterval(() => closeAlert(), 5000);
+  const closeAlert = () => {
+    let alerts = Array.from(document.getElementsByClassName('alert'));
+    alerts.forEach((alert) => {
+      alert.remove();
+    });
+    clearInterval(interval);
+  };
+
   return (
     <div
       className={
